add some fixes

+++ backend/utils/categories.go
@@ -2,7 +2,7 @@ package utils
 
 func CheckPOSTCategories(categories []any) bool {
 	for _, category := range categories {
-		_, ok := category.(int)
+		_, ok := category.(float64)
 		if !ok {
 			return false
 		}
